Refactor: Move EnterUsernameViewModel to its own file
This commit moves the `EnterUsernameViewModel` from `UsernameActivity.kt` to its own file `EnterUsernameViewModel.kt` within the `presentation.managers` package. This improves code organization and separation of concerns.
